Start implementing purerl FFI
The first thing we need is an FFI file.

Each exported function has to specify its arity.
it's a bit tricky since the PS imports most of the functions as FnX.

The way those are implemented in the compiler is as a nullary function
that returns an uncurried function of the appropriate arity.

For instance:

```PureScript
module Foo where

foriegn import foo :: Fn2 Int String Boolean
```

corresponds to

```Erlang
-module(foo@foreign).
-export([foo/0]).

foo() ->
    fun (N, Str) ->
        false
    end.
```

For now, we just return the atom of the name of the function.
It should help us pinpoint what we've not implemented yet.
As we implement behavior, we'll flesh out things more.

We start with the general pattern of an RPC loop.
We'll modify this as we go.

We also comment out all of the tests so we can run the suite.
We can start uncommenting the suite to guide implementation,
ala Test-Driven Development.