Implement `makeVar/1` and `'_tryReadVar'/0`
First we start by uncommenting the test to show us what we're missing.

It says that we've not implemented `makeVar/1`.
To implement `makeVar/1`,
we return a nullary function representing the `AVarEff eff (AVar a)`.
In this function, we spawn a process that invokes `filled/1` with
the given value.

`filled/1` represents one of the `AVarStatus a`s.
We should also have a `killed/1` and `empty/0` eventually.

We `receive` in `filled/1` until we get a message we can handle.
When we get a `tryRead` message, we respond with our value in a `Just`.

`rpc/2` takes care of all of the book keeping of ensuring the
messages are routed properly.
So we can write a slightly clearer version of this logic.