Implement correct behavior for `empty/2` `filled/2`
`AVar a` differs from `MVar a` in that `AVar a` is based around
callbacks.

The following code would block for an `MVar a`:

```Haskell
foo = do
  mVar <- newMVar 0
  putMVar mVar 1
  takeMVar mVar
```

Whereas its equvialent in PureScript does not block,
but queues up callbacks.

```PureScript
foo = do
  aVar <- newAVar 0
  putAVar 1 aVar \_ -> log "put"
  takeAVar aVar \_ -> log "take"
```

In order to handle the callbacks properly,
we keep track of a couple of queues.

`empty/2`s have a queue for `read`ing and one for `take`ing.
Reads always run to completion as soon as we have a value.
Takes run one at a time when we get a value.

`filled/2`s have one queue for `put`ing.